src/test/java/com/company/officecommute/service/team/TeamServiceTest.java
package com.company.officecommute.service.team;

import com.company.officecommute.domain.team.Team;
import com.company.officecommute.repository.team.TeamRepository;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.BDDMockito;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;

import static org.assertj.core.api.Assertions.assertThat;
import static org.mockito.ArgumentMatchers.any;

@ExtendWith(MockitoExtension.class)
class TeamServiceTest {

    @InjectMocks
    private TeamService teamService;

    @Mock
    private TeamRepository teamRepository;

    @Test
    void testRegisterTeam() {
        Team team = Teams.createTeamWithTeamName("ATeam");
        BDDMockito.given(teamRepository.save(any(Team.class))).willReturn(team);

        Long teamId = teamService.registerTeam("ATeam");

        assertThat(teamId).isEqualTo(1L);
    }

}
